requirements.txt
fastapi
uvicorn
pydantic
python-multipart
ultralytics
easyocr
opencv-python-headless
pyzbar
numpy
requests
Pillow